fg-sbir-runnable-demo/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24
pyyaml>=6.0.1
tqdm>=4.66
scikit-learn>=1.3
Pillow>=9.0
